moves most of the logic for actually signing to signeeService
