+++ src/Altinn.App.Core/Extensions/ServiceCollectionExtensions.cs
@@ -303,6 +303,7 @@ public static class ServiceCollectionExtensions
         services.AddHttpClient<ISignClient, SignClient>();
         services.AddTransient<ISigningDelegationService, SigningDelegationService>();
         services.AddTransient<ISigningNotificationService, SigningNotificationService>();
+        services.AddTransient<ISigningCorrespondenceService, SigningCorrespondenceService>();
         services.AddTransient<ISigningService, SigningService>();
     }
