+++ test/Altinn.App.Core.Tests/Features/Signing/SigningServiceTests.cs
@@ -1,15 +1,21 @@
 ﻿using System.Text;
 using System.Text.Json;
+using Altinn.App.Core.Configuration;
 using Altinn.App.Core.Features;
 using Altinn.App.Core.Features.Signing;
 using Altinn.App.Core.Features.Signing.Interfaces;
 using Altinn.App.Core.Features.Signing.Models;
+using Altinn.App.Core.Internal.App;
 using Altinn.App.Core.Internal.Process.Elements.AltinnExtensionProperties;
+using Altinn.App.Core.Internal.Profile;
 using Altinn.App.Core.Internal.Registers;
+using Altinn.App.Core.Internal.Sign;
 using Altinn.App.Core.Models;
 using Altinn.Platform.Register.Models;
 using Altinn.Platform.Storage.Interface.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
 using Moq;
 
 namespace Altinn.App.Core.Tests.Features.Signing;
@@ -23,6 +29,13 @@ public class SigningServiceTests
     private readonly Mock<ISigningNotificationService> _signingNotificationService = new(MockBehavior.Strict);
     private readonly Mock<ISigneeProvider> _signeeProvider = new(MockBehavior.Strict);
     private readonly Mock<ILogger<SigningService>> _logger = new(MockBehavior.Strict);
+    private readonly Mock<IAppMetadata> _appMetadata = new(MockBehavior.Strict);
+    private readonly Mock<IHttpContextAccessor> _httpContextAccessor = new(MockBehavior.Strict);
+    private readonly Mock<ISignClient> _signClient = new(MockBehavior.Strict);
+    private readonly Mock<ISigningCorrespondenceService> _signingCorrespondenceService = new(MockBehavior.Strict);
+    private readonly Mock<IProfileClient> _profileClient = new(MockBehavior.Strict);
+    private readonly Mock<IAltinnPartyClient> _altinnPartyClientService = new(MockBehavior.Strict);
+    private readonly Mock<IOptions<GeneralSettings>> _settings = new();
 
     public SigningServiceTests()
     {
@@ -31,6 +44,13 @@ public class SigningServiceTests
             _signingDelegationService.Object,
             _signingNotificationService.Object,
             [_signeeProvider.Object],
+            _appMetadata.Object,
+            _httpContextAccessor.Object,
+            _signClient.Object,
+            _signingCorrespondenceService.Object,
+            _profileClient.Object,
+            _altinnPartyClientService.Object,
+            _settings.Object,
             _logger.Object
         );
     }
@@ -90,14 +110,17 @@ public class SigningServiceTests
 
         var signDocumentWithMatchingSignatureContext = new SignDocument
         {
-            SigneeInfo = new Signee { OrganisationNumber = signeeState.First().Party.OrgNumber },
+            SigneeInfo = new Platform.Storage.Interface.Models.Signee
+            {
+                OrganisationNumber = signeeState.First().Party.OrgNumber,
+            },
         };
 
         var person = new Person { SSN = "12345678910", Name = "A person" };
 
         var signDocumentWithoutMatchingSignatureContext = new SignDocument
         {
-            SigneeInfo = new Signee { PersonNumber = person.SSN },
+            SigneeInfo = new Platform.Storage.Interface.Models.Signee { PersonNumber = person.SSN },
         };
 
         cachedInstanceMutator.Setup(x => x.Instance).Returns(instance);
